Logind

+++ src/locker.rs
@@ -24,8 +24,10 @@ use std::{
     collections::HashMap,
     ffi::{CStr, CString},
     fs,
+    os::fd::OwnedFd,
     path::Path,
     process,
+    sync::Arc,
 };
 use tokio::{sync::mpsc, task, time};
 use wayland_client::{protocol::wl_output::WlOutput, Proxy};
@@ -195,24 +197,36 @@ pub enum Message {
     SessionLockEvent(SessionLockEvent),
     Channel(mpsc::Sender<String>),
     BackgroundState(cosmic_bg_config::state::State),
+    Inhibit(Arc<OwnedFd>),
     NetworkIcon(Option<&'static str>),
     PowerInfo(Option<(String, f64)>),
     Prompt(String, bool, Option<String>),
     Submit,
     Suspend,
     Error(String),
-    Exit,
+    Lock,
+    Unlock,
+}
+
+#[derive(Clone, Debug)]
+enum State {
+    Locking,
+    Locked,
+    Unlocking,
+    Unlocked,
 }
 
 /// The [`App`] stores application-specific state.
 pub struct App {
     core: Core,
     flags: Flags,
+    state: State,
     surface_ids: HashMap<WlOutput, SurfaceId>,
     active_surface_id_opt: Option<SurfaceId>,
     surface_images: HashMap<SurfaceId, widget::image::Handle>,
     surface_names: HashMap<SurfaceId, String>,
     text_input_ids: HashMap<SurfaceId, widget::Id>,
+    inhibit_opt: Option<Arc<OwnedFd>>,
     network_icon_opt: Option<&'static str>,
     power_info_opt: Option<(String, f64)>,
     value_tx_opt: Option<mpsc::Sender<String>>,
@@ -302,18 +316,26 @@ impl cosmic::Application for App {
             App {
                 core,
                 flags,
+                state: State::Unlocked,
                 surface_ids: HashMap::new(),
                 active_surface_id_opt: None,
                 surface_images: HashMap::new(),
                 surface_names: HashMap::new(),
                 text_input_ids: HashMap::new(),
+                inhibit_opt: None,
                 network_icon_opt: None,
                 power_info_opt: None,
                 value_tx_opt: None,
                 prompt_opt: None,
                 error_opt: None,
             },
-            lock(),
+            if cfg!(feature = "logind") {
+                // When logind feature is used, wait for lock signal
+                Command::none()
+            } else {
+                // When logind feature not used, lock immediately
+                lock()
+            },
         )
     }
 
@@ -359,10 +381,12 @@ impl cosmic::Application for App {
                         self.text_input_ids
                             .insert(surface_id, text_input_id.clone());
 
-                        return Command::batch([
-                            get_lock_surface(surface_id, output),
-                            widget::text_input::focus(text_input_id),
-                        ]);
+                        if matches!(self.state, State::Locked) {
+                            return Command::batch([
+                                get_lock_surface(surface_id, output),
+                                widget::text_input::focus(text_input_id),
+                            ]);
+                        }
                     }
                     OutputEvent::Removed => {
                         log::info!("output {}: removed", output.id());
@@ -371,15 +395,17 @@ impl cosmic::Application for App {
                                 self.surface_images.remove(&surface_id);
                                 self.surface_names.remove(&surface_id);
                                 self.text_input_ids.remove(&surface_id);
-                                return destroy_lock_surface(surface_id);
+                                if matches!(self.state, State::Locked) {
+                                    return destroy_lock_surface(surface_id);
+                                }
                             }
                             None => {
                                 log::warn!("output {}: no surface found", output.id());
                             }
                         }
                     }
-                    OutputEvent::InfoUpdate(output_info) => {
-                        log::info!("output {}: info update {:#?}", output.id(), output_info);
+                    OutputEvent::InfoUpdate(_output_info) => {
+                        log::info!("output {}: info update", output.id());
                     }
                 }
             }
@@ -391,10 +417,29 @@ impl cosmic::Application for App {
                         return widget::text_input::focus(text_input_id.clone());
                     }
                 }
-                SessionLockEvent::Unlocked => {
-                    //TODO: cleaner method to exit?
-                    process::exit(0);
+                SessionLockEvent::Locked => {
+                    log::info!("session locked");
+                    self.state = State::Locked;
+                    // Allow suspend
+                    self.inhibit_opt = None;
+                    let mut commands = Vec::with_capacity(self.surface_ids.len());
+                    for (output, surface_id) in self.surface_ids.iter() {
+                        commands.push(get_lock_surface(*surface_id, output.clone()));
+                    }
+                    return Command::batch(commands);
                 }
+                SessionLockEvent::Unlocked => {
+                    log::info!("session unlocked");
+                    self.state = State::Unlocked;
+                    if cfg!(feature = "logind") {
+                        // When using logind feature, stick around for more lock signals
+                    } else {
+                        // When not using logind feature, exit immediately after unlocking
+                        //TODO: cleaner method to exit?
+                        process::exit(0);
+                    }
+                }
+                //TODO: handle finished signal
                 _ => {}
             },
             Message::Channel(value_tx) => {
@@ -404,6 +449,9 @@ impl cosmic::Application for App {
                 self.flags.wallpapers = background_state.wallpapers;
                 self.surface_images.clear();
                 self.update_wallpapers();
+            }
+            Message::Inhibit(inhibit) => {
+                self.inhibit_opt = Some(inhibit);
             }
             Message::NetworkIcon(network_icon_opt) => {
                 self.network_icon_opt = network_icon_opt;
@@ -426,6 +474,8 @@ impl cosmic::Application for App {
                 Some((_prompt, _secret, value_opt)) => match value_opt {
                     Some(value) => match self.value_tx_opt.take() {
                         Some(value_tx) => {
+                            // Clear errors
+                            self.error_opt = None;
                             return Command::perform(
                                 async move {
                                     value_tx.send(value).await.unwrap();
@@ -458,17 +508,47 @@ impl cosmic::Application for App {
             Message::Error(error) => {
                 self.error_opt = Some(error);
             }
-            Message::Exit => {
-                let mut commands = Vec::new();
-                for (_output, surface_id) in self.surface_ids.drain() {
-                    self.surface_images.remove(&surface_id);
-                    self.surface_names.remove(&surface_id);
-                    self.text_input_ids.remove(&surface_id);
-                    commands.push(destroy_lock_surface(surface_id));
+            Message::Lock => match self.state {
+                State::Unlocked => {
+                    log::info!("session locking");
+                    self.state = State::Locking;
+                    // Clear errors
+                    self.error_opt = None;
+                    // Clear value_tx
+                    self.value_tx_opt = None;
+                    return lock();
                 }
-                commands.push(unlock());
-                // Wait to exit until `Unlocked` event, when server has processed unlock
-                return Command::batch(commands);
+                State::Unlocking => {
+                    log::info!("session still unlocking");
+                }
+                State::Locking | State::Locked => {
+                    log::info!("session already locking or locked");
+                }
+            },
+            Message::Unlock => {
+                match self.state {
+                    State::Locked => {
+                        log::info!("sessing unlocking");
+                        self.state = State::Unlocking;
+                        // Clear errors
+                        self.error_opt = None;
+                        // Clear value_tx
+                        self.value_tx_opt = None;
+                        let mut commands = Vec::with_capacity(self.surface_ids.len() + 1);
+                        for (_output, surface_id) in self.surface_ids.iter() {
+                            commands.push(destroy_lock_surface(*surface_id));
+                        }
+                        commands.push(unlock());
+                        // Wait to exit until `Unlocked` event, when server has processed unlock
+                        return Command::batch(commands);
+                    }
+                    State::Locking => {
+                        log::info!("session still locking");
+                    }
+                    State::Unlocking | State::Unlocked => {
+                        log::info!("session already unlocking or unlocked");
+                    }
+                }
             }
         }
         Command::none()
@@ -655,42 +735,23 @@ impl cosmic::Application for App {
     }
 
     fn subscription(&self) -> Subscription<Self::Message> {
-        struct BackgroundSubscription;
-        struct HeartbeatSubscription;
-        struct PamSubscription;
+        let mut subscriptions = Vec::with_capacity(7);
 
-        //TODO: just use one vec for all subscriptions
-        let mut extra_suscriptions = Vec::with_capacity(2);
-
-        #[cfg(feature = "networkmanager")]
-        {
-            extra_suscriptions.push(
-                crate::networkmanager::subscription()
-                    .map(|icon_opt| Message::NetworkIcon(icon_opt)),
-            );
-        }
-
-        #[cfg(feature = "upower")]
-        {
-            extra_suscriptions
-                .push(crate::upower::subscription().map(|info_opt| Message::PowerInfo(info_opt)));
-        }
-
-        //TODO: how to avoid cloning this on every time subscription is called?
-        let username = self.flags.current_user.name.clone();
-        Subscription::batch([
-            event::listen_with(|event, _| match event {
-                iced::Event::PlatformSpecific(iced::event::PlatformSpecific::Wayland(
-                    wayland_event,
-                )) => match wayland_event {
-                    WaylandEvent::Output(output_event, output) => {
-                        Some(Message::OutputEvent(output_event, output))
-                    }
-                    WaylandEvent::SessionLock(evt) => Some(Message::SessionLockEvent(evt)),
-                    _ => None,
-                },
+        subscriptions.push(event::listen_with(|event, _| match event {
+            iced::Event::PlatformSpecific(iced::event::PlatformSpecific::Wayland(
+                wayland_event,
+            )) => match wayland_event {
+                WaylandEvent::Output(output_event, output) => {
+                    Some(Message::OutputEvent(output_event, output))
+                }
+                WaylandEvent::SessionLock(evt) => Some(Message::SessionLockEvent(evt)),
                 _ => None,
-            }),
+            },
+            _ => None,
+        }));
+
+        struct BackgroundSubscription;
+        subscriptions.push(
             cosmic_config::config_state_subscription(
                 TypeId::of::<BackgroundSubscription>(),
                 cosmic_bg_config::NAME.into(),
@@ -702,7 +763,11 @@ impl cosmic::Application for App {
                 }
                 Message::BackgroundState(res.config)
             }),
-            subscription::channel(
+        );
+
+        if matches!(self.state, State::Locked) {
+            struct HeartbeatSubscription;
+            subscriptions.push(subscription::channel(
                 TypeId::of::<HeartbeatSubscription>(),
                 16,
                 |mut msg_tx| async move {
@@ -713,8 +778,12 @@ impl cosmic::Application for App {
                         time::sleep(time::Duration::new(1, 0)).await;
                     }
                 },
-            ),
-            subscription::channel(
+            ));
+
+            struct PamSubscription;
+            //TODO: how to avoid cloning this on every time subscription is called?
+            let username = self.flags.current_user.name.clone();
+            subscriptions.push(subscription::channel(
                 TypeId::of::<PamSubscription>(),
                 16,
                 |mut msg_tx| async move {
@@ -735,7 +804,7 @@ impl cosmic::Application for App {
                         match pam_res {
                             Ok(()) => {
                                 log::info!("successfully authenticated");
-                                msg_tx.send(Message::Exit).await.unwrap();
+                                msg_tx.send(Message::Unlock).await.unwrap();
                                 break;
                             }
                             Err(err) => {
@@ -749,8 +818,28 @@ impl cosmic::Application for App {
                         time::sleep(time::Duration::new(60, 0)).await;
                     }
                 },
-            ),
-            Subscription::batch(extra_suscriptions),
-        ])
+            ));
+        }
+
+        #[cfg(feature = "logind")]
+        {
+            subscriptions.push(crate::logind::subscription());
+        }
+
+        #[cfg(feature = "networkmanager")]
+        {
+            subscriptions.push(
+                crate::networkmanager::subscription()
+                    .map(|icon_opt| Message::NetworkIcon(icon_opt)),
+            );
+        }
+
+        #[cfg(feature = "upower")]
+        {
+            subscriptions
+                .push(crate::upower::subscription().map(|info_opt| Message::PowerInfo(info_opt)));
+        }
+
+        Subscription::batch(subscriptions)
     }
 }
